python_backend/requirements.txt
# Flask Backend Dependencies for Eye Disease Detection

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# Deep Learning
tensorflow>=2.13.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0

# Utilities
python-dotenv==1.0.0
gunicorn==21.2.0  # For production deployment
opencv-python-headless>=4.8.0  # For image validation
